Commit 40

+++ pipelines/abalone/.ipynb_checkpoints/preprocess-checkpoint.py
@@ -77,7 +77,7 @@ if __name__ == "__main__":
      ]
     ]
     
-    df['medianHouseValue'] = df['medianHouseValue'].astype('int')
+    df['medianHouseValue'] = df['medianHouseValue'].astype(float)
     Y = df[["medianHouseValue"]] / 100000
 
     x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.33)
